New stamina efficiency logic to replace stamina reduction logic (basically the same thing with math little bit different and a bit more generified)

+++ src/main/java/tictim/paraglider/client/render/InGameStaminaWheelRenderer.java
@@ -35,7 +35,7 @@ public class InGameStaminaWheelRenderer extends StaminaWheelRenderer {
 		double stamina = s.stamina();
 
 		Movement movement = Movement.get(player);
-		int staminaDelta = movement.staminaDelta();
+		double staminaDelta = movement.staminaDelta();
 
 		wheel.setProperties(stamina, maxStamina);
